+++ src/Parquet.Test/Serialisation/ParquetSerializerTest.cs
@@ -22,16 +22,12 @@ namespace Parquet.Test.Serialisation;
 /// </summary>
 public class ParquetSerializerTest : TestBase {
 
-    private async Task Compare<T>(List<T> input, bool asJson = false, string? saveAsFile = null, bool useAsync = false)
+    private async Task Compare<T>(List<T> input, bool asJson = false, string? saveAsFile = null)
         where T : class, new() {
 
         // serialize to parquet
         using var ms = new MemoryStream();
-        if(useAsync) {
-            await ParquetSerializer.SerializeAsync(input.ToAsyncEnumerable(), ms);
-        } else {
-            await ParquetSerializer.SerializeAsync(input, ms);
-        }
+        await ParquetSerializer.SerializeAsync(input, ms);
 
         if(saveAsFile != null) {
             await System.IO.File.WriteAllBytesAsync(saveAsFile, ms.ToArray());
@@ -54,7 +50,7 @@ public class ParquetSerializerTest : TestBase {
 
         // serialize to parquet
         using var ms = new MemoryStream();
-        await ParquetSerializer.SerializeAsync(typeof(TSchema).GetParquetSchema(true), input, ms);
+        await ParquetSerializer.SerializeAsync(input, typeof(TSchema).GetParquetSchema(true), ms);
 
         if(writeTestFile != null) {
             System.IO.File.WriteAllBytes(writeTestFile, ms.ToArray());
@@ -80,10 +76,8 @@ public class ParquetSerializerTest : TestBase {
         public Guid ExternalId { get; set; }
     }
 
-    [Theory]
-    [InlineData(false)]
-    [InlineData(true)]
-    public async Task Atomics_Simplest_Serde(bool useAsync) {
+    [Fact]
+    public async Task Atomics_Simplest_Serde() {
 
         var data = Enumerable.Range(0, 1_000).Select(i => new Record {
             Timestamp = DateTime.UtcNow.AddSeconds(i),
@@ -92,7 +86,7 @@ public class ParquetSerializerTest : TestBase {
             ExternalId = Guid.NewGuid()
         }).ToList();
 
-        await Compare(data, useAsync: useAsync);
+        await Compare(data);
     }
 
     class TimespanRecord {
@@ -861,10 +855,8 @@ public class ParquetSerializerTest : TestBase {
         }
     }
 
-    [Theory]
-    [InlineData(false)]
-    [InlineData(true)]
-    public async Task Specify_row_group_size(bool useAsync) {
+    [Fact]
+    public async Task Specify_row_group_size() {
         var data = Enumerable.Range(0, 100).Select(i => new Record {
             Timestamp = DateTime.UtcNow.AddSeconds(i),
             EventName = i % 2 == 0 ? "on" : "off",
@@ -873,11 +865,7 @@ public class ParquetSerializerTest : TestBase {
 
         using var ms = new MemoryStream();
         var options = new ParquetSerializerOptions { RowGroupSize = 20 };
-        if(useAsync) {
-            await ParquetSerializer.SerializeAsync(data.ToAsyncEnumerable(), ms, options);
-        } else {
-            await ParquetSerializer.SerializeAsync(data, ms, options);
-        }
+        await ParquetSerializer.SerializeAsync(data, ms, options);
 
         // validate we have 5 row groups in the resulting file
         ms.Position = 0;
@@ -902,38 +890,11 @@ public class ParquetSerializerTest : TestBase {
 
         ms.Position = 0;
         for(int i = 0; i < records; i += rowGroupSize) {
-            IList<Record> data2 = await ParquetSerializer.DeserializeAsync<Record>(ms, i / rowGroupSize);
+            IList<Record> data2 = await ParquetSerializer.DeserializeAsync<Record>(ms, rowGroupIndex: i / rowGroupSize);
             Assert.Equivalent(data.Skip(i).Take(rowGroupSize), data2);
         }
     }
 
-    [Theory]
-    [InlineData(false)]
-    [InlineData(true)]
-    public async Task Deserialize_as_async_enumerable(bool useAsync) {
-        DateTime now = DateTime.UtcNow;
-        int records = 100;
-        int rowGroupSize = 20;
-
-        var data = Enumerable.Range(0, records).Select(i => new Record {
-            Timestamp = now.AddSeconds(i),
-            EventName = i % 2 == 0 ? "on" : "off",
-            MeterValue = i
-        }).ToList();
-
-        using var ms = new MemoryStream();
-        var options = new ParquetSerializerOptions { RowGroupSize = rowGroupSize };
-        if(useAsync) {
-            await ParquetSerializer.SerializeAsync(data.ToAsyncEnumerable(), ms, options);
-        } else {
-            await ParquetSerializer.SerializeAsync(data, ms, options);
-        }
-
-        ms.Position = 0;
-        Record[] data2 = await ParquetSerializer.DeserializeAllAsync<Record>(ms).ToArrayAsync();
-
-        Assert.Equivalent(data, data2);
-    }
 
     [Fact]
     public async Task Specify_row_group_size_too_small() {
@@ -1061,7 +1022,7 @@ public class ParquetSerializerTest : TestBase {
         await ParquetSerializer.SerializeAsync(data, ms);
 
         ms.Position = 0;
-        IList<InterfaceImpl> data2 = await ParquetSerializer.DeserializeAllAsync<InterfaceImpl>(ms).ToArrayAsync();
+        IList<InterfaceImpl> data2 = await ParquetSerializer.DeserializeAsync<InterfaceImpl>(ms);
 
         Assert.Equivalent(data, data2);
 
@@ -1078,7 +1039,7 @@ public class ParquetSerializerTest : TestBase {
         await ParquetSerializer.SerializeAsync(data, ms);
 
         ms.Position = 0;
-        IList<ConcreteRootInterfaceImpl> data2 = await ParquetSerializer.DeserializeAllAsync<ConcreteRootInterfaceImpl>(ms).ToArrayAsync();
+        IList<ConcreteRootInterfaceImpl> data2 = await ParquetSerializer.DeserializeAsync<ConcreteRootInterfaceImpl>(ms);
 
         Assert.Equivalent(data, data2);
     }
@@ -1264,33 +1225,6 @@ public class ParquetSerializerTest : TestBase {
         public Dictionary<DateTime, double> DateTimeDoubleDict { get; set; } = new();
     }
 
-    [Fact]
-    public async Task E2EPreInitializedDict() {
-        using var ms = new MemoryStream();
-
-        var testData = new ClassWithPreinitializedDict {
-            DateTimeDoubleDict = new Dictionary<DateTime, double> {
-                { new DateTime(2021, 1, 1), 1.1 },
-                { new DateTime(2021, 1, 2), 2.2 },
-                { new DateTime(2021, 1, 3), 3.3 }
-            }
-        };
-
-        await using(ParquetWriter writer = await ParquetWriter.CreateAsync(typeof(ClassWithPreinitializedDict).GetParquetSchema(true), ms)) {
-            await ParquetSerializer.SerializeRowGroupAsync(writer, new[] { testData }, default);
-        }
-        ms.Position = 0;
-
-        await using ParquetReader reader = await ParquetReader.CreateAsync(ms);
-        ParquetSchema schema = reader.Schema;
-        using ParquetRowGroupReader rg = reader.OpenRowGroupReader(0);
-        var buffer = new List<ClassWithPreinitializedDict>();
-        await ParquetSerializer.DeserializeRowGroupAsync(rg, schema, buffer);
-
-        Assert.Equal(testData.DateTimeDoubleDict.Count, buffer[0]?.DateTimeDoubleDict?.Count);
-        Assert.Equal(testData.DateTimeDoubleDict[new DateTime(2021, 1, 1)], buffer[0]?.DateTimeDoubleDict?[new DateTime(2021, 1, 1)]);
-    }
-
     public record Record1 {
         public int Id { get; set; }
     }
@@ -1305,15 +1239,5 @@ public class ParquetSerializerTest : TestBase {
         ms.Position = 0;
         IList<Record1> data = await ParquetSerializer.DeserializeAsync<Record1>(ms);
         Assert.Empty(data);
-
-        // try with DeserializeAllAsync too
-        ms.Position = 0;
-        int count = 0;
-        await foreach(Record1 record in ParquetSerializer.DeserializeAllAsync<Record1>(ms)) {
-            // just iterate through all records
-            count += 1;
-        }
-        Assert.Equal(count, data.Count);
     }
-
 }